+++ tests/e2e-tests/src/main/kotlin/api_models/CreatePrismDidRequest.kt
@@ -1,5 +1,5 @@
 package api_models
 
-data class CreateManagedDidRequest(
+data class CreatePrismDidRequest(
     val documentTemplate: DocumentTemplate,
 )